+++ hybris/egl/platforms/hwcomposer/hwcomposer_window.h
@@ -31,7 +31,7 @@ protected:
     HWComposerNativeWindowBuffer(unsigned int width,
                                  unsigned int height,
                                  unsigned int format,
-                                 unsigned int usage) ;
+                                 uint64_t usage) ;
    virtual ~HWComposerNativeWindowBuffer() ;
 
 protected:
